+++ tests/dummy/mirage/fixtures/view.js
@@ -694,14 +694,12 @@ export default [
           rows: [
             [
               {
-                model: 'enumExample',
-                renderer: 'select'
+                model: 'enumExample'
               }
             ],
             [
               {
-                model: 'queryExample',
-                renderer: 'select'
+                model: 'queryExample'
               }
             ],
             [
@@ -734,14 +732,12 @@ export default [
           rows: [
             [
               {
-                model: 'enumExample',
-                renderer: 'select'
+                model: 'enumExample'
               }
             ],
             [
               {
                 model: 'queryExample',
-                renderer: 'select',
                 writeTransforms: [
                   {
                     object: {
@@ -897,10 +893,7 @@ export default [
               model: 'lastNameAtBirth'
             }],
             [{model: 'dateOfBirth'}],
-            [{
-              model: 'countryOfBirth',
-              renderer: 'select'
-            }],
+            [{model: 'countryOfBirth'}],
             [{model: 'stateOfBirth'}]
           ]
         },
@@ -911,10 +904,7 @@ export default [
             [{model: 'address'}],
             [{model: 'city'}],
             [{model: 'state'}],
-            [{
-              model: 'country',
-              renderer: 'select'
-            }],
+            [{model: 'country'}],
             [{model: 'zipCode'}]
           ]
         },
@@ -940,10 +930,7 @@ export default [
             [{model: 'middleName'}],
             [{model: 'lastName'}],
             [{model: 'stateOfBirth'}],
-            [{
-              model: 'countryOfBirth',
-              renderer: 'select'
-            }]
+            [{model: 'countryOfBirth'}]
           ]
         },
         {
@@ -953,10 +940,7 @@ export default [
             [{model: 'middleName'}],
             [{model: 'lastName'}],
             [{model: 'stateOfBirth'}],
-            [{
-              model: 'countryOfBirth',
-              renderer: 'select'
-            }]
+            [{model: 'countryOfBirth'}]
           ]
         }
       ],
@@ -987,8 +971,7 @@ export default [
           rows: [
             [
               {
-                model: 'tagType',
-                renderer: 'select'
+                model: 'tagType'
               }
             ],
             [
@@ -1025,8 +1008,7 @@ export default [
           rows: [
             [
               {
-                model: 'tagType',
-                renderer: 'select'
+                model: 'tagType'
               }
             ],
             [
@@ -1063,8 +1045,7 @@ export default [
           rows: [
             [
               {
-                model: 'tagType',
-                renderer: 'select'
+                model: 'tagType'
               }
             ],
             [