+++ tests/unit/components/input-wrapper-test.js
@@ -1,5 +1,6 @@
+import {expect} from 'chai'
 import {describeComponent} from 'ember-mocha'
-import {beforeEach} from 'mocha'
+import {beforeEach, describe, it} from 'mocha'
 import {PropTypes} from 'ember-prop-types'
 import {validatePropTypes} from 'dummy/tests/helpers/template'
 
@@ -10,9 +11,12 @@ describeComponent(
     unit: true
   },
   function () {
+    let component
+
     beforeEach(function () {
-      this.subject({
+      component = this.subject({
         bunsenId: 'foo',
+
         bunsenStore: Ember.Object.create({})
       })
     })
@@ -33,6 +37,32 @@ describeComponent(
         PropTypes.object,
         PropTypes.string
       ])
+    })
+
+    describe('inputName', function () {
+      describe('when model specifies enum', function () {
+        beforeEach(function () {
+          component.set('bunsenModel', {
+            enum: ['a', 'b', 'c']
+          })
+        })
+
+        it('returns select renderer', function () {
+          expect(component.get('inputName')).to.equal('frost-bunsen-input-select')
+        })
+      })
+
+      describe('when model specifies modelType', function () {
+        beforeEach(function () {
+          component.set('bunsenModel', {
+            modelType: 'bar'
+          })
+        })
+
+        it('returns select renderer', function () {
+          expect(component.get('inputName')).to.equal('frost-bunsen-input-select')
+        })
+      })
     })
   }
 )