+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 6.6.1
+
+* **Fixed** code to use `Ember.RSVP` instead of native `Promise`'s to make the Ember run loop aware. This fixes failing acceptance tests in a consuming application.
+
+# 6.6.0
+
+* Made it so the `select` renderer is automatically used when you specify the `enum` or `modelType` property for a model attribute.
+
+# 6.5.0
+
+* **Added** new `textarea` renderer than wraps the `frost-textarea` component from `ember-frost-core`.
+
 # 6.4.1
 
 * **Added** `ember-disable-prototype-extensions` to ensure code works with prototype extensions disabled.